+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-06-21
+
 ### Changed
 
 - `pool_size` options of `Mxpanel.Batcher` increased from `3` to `5`.
@@ -30,5 +32,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Mxpanel.track_many/2`.
 - `Mxpanel.track_later/2`.
 
-[Unreleased]: https://github.com/thiamsantos/mxpanel/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/thiamsantos/mxpanel/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/thiamsantos/mxpanel/releases/tag/v0.2.0
 [0.1.0]: https://github.com/thiamsantos/mxpanel/releases/tag/v0.1.0